Changes EIP-1 wording to focus on technicals not community sentiment. (#1224)
* Changes EIP-1 wording to focus on technicals not community sentiment.

All of the recent changes to the EIP process have been made to ensure that the EIP process is a technical one, and not one of sentiment analysis.  There is a lot of discussion going on with regards to how we can improve the process and get valid community sentiment analysis pre-fork, but there doesn't exist a complete solution yet at this time (just proposals).

It appears that the Last Call PR introduced sentiment analysis into the process, which I do not believe was intended.  From my recollection of the discussions around the Last Call stuff, the goal wasn't to fundamentally change how governance works, but rather to ensure that EIPs don't get stuck indefinitely in limbo.

This change simply removes the sentiment analysis wording from the process and makes it more clear that the EIP process is about gauging technical feasibility, not making judgement calls as to whether or not a thing is a good idea or not.

* Adds Active state