Cleans up EIP-55 (#3092)
* Cleans up EIP-55

This EIP was written in a time when we were using EIPs for hardfork coordination and we weren't worried too much about link rot.  However, times have changed and maintaining EIPs like this is becoming more and more of a burden.  In this case, the specification is what is important, not who implements it.  Also, editors do not want to maintain external links to things that don't add *significant* value.

* Update eip-55.md